Refactored (java-application): Created for Database to two different Interfaces
