+++ server/src/main/java/Database/DatabaseModule.java
@@ -1,21 +1,18 @@
 package Database;
 
 import Config.Config;
+import Database.Impl.DatabaseImpl;
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 import com.google.inject.Singleton;
 
 public class DatabaseModule extends AbstractModule
 {
-//    @Provides @Singleton
-//    static DatabaseProvider provideDatabaseController(Config config)
-//    {
-//        return new DatabaseProvider(config);
-//    }
-
     @Override
     protected void configure()
     {
-       // this.bind(DatabaseProvider.class).to(DatabaseProvider.class).in(Singleton.class);
+        this.bind(DatabaseImpl.class).in(Singleton.class);
+        this.bind(Database.class).to(DatabaseImpl.class);
+        this.bind(DatabaseService.class).to(DatabaseImpl.class);
     }
 }
