+++ server/src/main/java/Database/DatabaseService.java
@@ -1,0 +1,13 @@
+package Database;
+
+import com.zaxxer.hikari.HikariDataSource;
+import org.jooq.DSLContext;
+
+import java.sql.Connection;
+import java.sql.SQLException;
+
+public interface DatabaseService
+{
+    public void start();
+    public void shutdown();
+}
